[FIX] account_consolidation: no delete other companies mapped accounts
When importing data via a file with only one company selected, if there is no mapped account
for a consolidation account in the file the system will empty the mapped account for all
the companies and not just the one selected.

Now the import does not take actions when empty mapped accounts.

opw-2449108

closes odoo/enterprise#18094

X-original-commit: 6fe6fac813331b60f03fd41cf8c18ca7aaab0c17
Signed-off-by: agr-odoo <agr-odoo@users.noreply.github.com>