[FIX] planning: format allocated hours in popover
Previously, the allocated hours was formatted with Math.round for the
minutes, so if a slot had a duration of 7:59:31, the result would be
7:60 which is not an hour.

Reminder : In gantt popover we are not able to use widgets so we have to
format the values before the qweb rendering.

This commit uses the float_time formatter to format allocated_hours and
then pass it to renderer.

PR : #15921

task-2215216

Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>