[IMP] planning: add job title and conflict filters
This commit aims to add filters on job title and shifts in conflict to
planning and project_forecast.

=== Filters

We implement for slot in conflict a search method
for a simple reason : the creation or edition of an other slot may
affect the value of those computed fields.

To take account of those environnement changes, we have to define
_search methods.

PR enterprise : odoo#15921

task-2215216