[FIX] account_intrastat: set product origin country of AML
When validating a PO, the product's country is not saved on the account
move line. Therefore, this information is missing in intrastat report.

To reproduce the error:
(Need purchase. Use a BE company)
1. In Settings, enable "Intrastat"
2. Create a product P
    - Accounting > Country of Origin: France
3. Create a vendor V
    - Must be European but not Belgian
4. Create a PO
    - Vendor: V
    - Product: P
5. Set the received quantity
6. Create Bill, Confirm
7. Accounting > Reporting > Intrastat Report

Error: The "Origin Country" of the line is "QU". It should be "FR".

This fix suggests to convert the `onchange` method into a `compute` one
since:
- The body is the expected behaviour
- This is not the initial `_onchange_product_id`, so the overrides are
not lost

OPW-2474599

X-original-commit: ce4067b45040b54f75b43a6a52e6665dcb6b4da2