[FIX] web_studio: include default values in export
The export of custom modules does not contain the default values. This
leads to a second problem: if a model can be archived, since the default
value of `active` is not defined, the new records will be considered as
archived.

To reproduce the error:
1. Enable Studio mode
2. New App, New Model M
    - Enable the "Archiving" feature
3. On M's form, add a default value to "Name" field
4. Back on Odoo menu, still in Studio mode:
    - Customizations > Export > Save on your device
5. Open a new DB
6. In Studio mode:
    - Customizations > Import > Select .zip
7. Close Studio mode
8. Create a new M

Error: The default value of "Name" field is not present. Worst, the
"Archived" label is displayed on the form: the new record will be
archived.

The data export does not include the default values (`ir.default`). To
retrieve the data, the export is based on the `ir.model.data` table.
This is the reason why this fix suggests that `ir.model.data` should
include `ir.default` records that are created for fields used in custom
modules.

OPW-2476907

closes odoo/enterprise#18176

X-original-commit: cf82d5a6bd931be9fa258739063e0799c9d7c413
Signed-off-by: Adrien Widart <adwid@users.noreply.github.com>