[IMP] hr: improves field type from text to html
Replace text fields to html fields as we have our own 'OdooEditor'.
Indeed, it gives more options to users in the way they format their
content without weighting too much on the UI
(tools appear on demand and not by default).

In this commit we also add new compute text field for hr.appraisal.goal
we have to add this field to display description in kanban and need to apply
some slicing in it, which will be not possible in html field due to html tags.

Models -> Fields

1) approval.request -> reason
2) hr.appraisal.goal -> description
3) hr.salary.rule -> note
4) hr.rule.parameter -> description
5) hr.salary.rule.category -> note
6) hr.payroll.structure -> note
7) hr.referral.reward -> description

Task Id: 2499504

X-original-commit: 3c4cc838d10320285b5ea9c53d24a05378b6bbac