[IMP] account_reports, l10n_*_reports: tax_report introduce tax units
To handle multicompany in the tax report, it is now possible to define tax units. A tax unit is a group of companies for which the tax report is submitted together to the state.

At most one tax unit can be defined per country. When there is one available, an option is available on the tax report to enable of of its companies and display its content. When opening the report with all the companies of the unit in self.env.companies, the options is activated by default on the localized tax report. On generic tax report, a selector is available for all tax units available, whatever their country, none is selected by default unless there is only one and its companies are active.

Tax reports now only allow multicompany through the usage of tax units.

Task 2463889

closes odoo/enterprise#18234

Related: odoo/upgrade#2482
Related: odoo/odoo#70588
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>