[FIX] account_followup: Consider only pertinent move lines
Before this commit the subqueries did select any move line
linked to the partner to set the 'with_overdue_invoices'
followup status.
Therefore, lines related to payment were selected since
there was no where clause on account type, and it did
also select already reconciled lines.
With this commit, we make sure that only unreconciled
lines on receivable accounts are selected and that the
followup status is correct.
It actually restores all the condition that were defined
in the WHERE clause of the CTE that has been removed for
performance reasons.

closes odoo/enterprise#18155

X-original-commit: a860cfaabd28ac8229de7542803de5215b1ba3e2
Signed-off-by: William André (wan) <wan@odoo.com>