[FIX] project_forecast: fix shift visibility for project user
Purpose of the commit is, the visibility of the project should
no longer have an impact on the visibility of a shift.
The fact that the user is a project > admin or a project > user should
also no longer have an impact on the visibility of a shift.

So in this commit, keep the following logic:
- internal user = own published shifts
- planning user = all published shifts
- planning admin = all shifts without distinction

- we will show shifts for private projects even if the user doesn't
have the rights for the private project.
- if project is private and user clicks on the project or task m2o,
we raise the usual access right error and if user does not have right
for the project user then do not allow to open the form view from list
and form view by no_open.
- also there should not be any access error while sending mail or
viewing the private project shift by user which does not have right
for that private project so used sudo to fetch some details from
private project/task.
- also display the open shift of private project on portal user so
user as take those shifts.

PS: this is the removal of commit: https://github.com/odoo/enterprise/commit/084f1d171443c7e346a2782a47f40967270b6813
with partical modification.

closes odoo/enterprise#18143

Taskid: 2523446
Related: odoo/upgrade#2458
Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>