[FIX] iot: Deadlock on exception in IngenicoDriver
The exception handler in IngenicoDriver's run method would call
`disconnect()`, which deleted the thread from the iot_devices and
socket_devices global dict. When the terminal tried to start a new
connection because the keepalive interval has passed, a new connection
would be created by `SocketInterface`. However, since this new
connection would have the same key as the old one, `update_iot_devices`
in `Interface` would not detect that and wouldn't create a new
IngenicoDriver thread in response, resulting in a deadlock.

To fix this, when the IngenicoDriver decides it's going to end its own
thread, it'll close the socket and will delete iot_devices, but it
will leave the socket_devices entry alone. When the terminal attempts
to create a new connection `replace_socket_device` will get called in
the `SocketInterface` and it will have the ability to update
`_detected_devices` in `Interface` so that `update_iot_devices` will
detect there's a new connection and create a new IngenicoDriver thread.

closes odoo/enterprise#18267

X-original-commit: 63b56f284a2c96f34788efc1f5c797a4970e38f9
Related: odoo/odoo#70657
Signed-off-by: Quentin Lejeune (qle) <qle@odoo.com>