[FIX] account_reports: fix missing lines with multi-currencies
- Use a company using USD as currency
- In Accounting, activate "Multi-Currencies"
- Configure currency rates as followed:
  * EUR - 01/01/2021 - 1.0
  * USD - 01/01/2021 - 1.2834
  * USD - 01/05/2021 - 1.38
- Create an invoice in EUR with an invoice line of 100€ and an invoice date
  between 01/01/2021 and 01/05/2021 (i.e. 01/03/2021)
- Create a payment for 99€ and keep the remaining 1€ open with payment date
  set at current date (i.e. 06/05/2021)
- Go to Accounting > Reporting > Aged Receivable
The invoice with the remaining unpaid 1€ does not appear in the report.
The payment is displayed instead with a negative value.

This is due to the difference of currencies rates between the invoice date
and the payment date.
Even though the amount of the payment (99€) is lower than the amount of the
invoice (100€), the converted amount in USD is greater.
As the report only takes into account converted values, it cannot handle
these cases with different currency rates.

To handle these cases, the report will use the currency amount (in the
original currency) when all account move lines use the same currency that
is different from the company currency.

opw-2514319

closes odoo/enterprise#18363

X-original-commit: e5ce657a8ef7e2fd6fa2633434d969775770026e
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Anh Thao PHAM <kitan191@users.noreply.github.com>