[FIX] hr_appraisal: confirm has link in email
Doing "1 Employee" -> "Request Appraisal" send a mail with a button
linking to the appraisal.

Doing "1 Appraisal" -> "Confirm" sends a mail without any link to the
appraisal.

In this changeset, we provides values for `recipient_users` and `url`
that are used in the default mail template to show the button.

opw-2486362

closes odoo/enterprise#18191

X-original-commit: dff8983cdf4abbb874279bf68c94eda48ab59da4
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>