[FIX] helpdesk(_*): prevent some data copying while duplicating ticket
Currently, after activating the sales services,
When we duplicating the ticket having records attached for return, coupons
and credit notes are also copied with the new tickets.

So in this commit, do not copy the some field while ducpliating the
ticket so it does not show on new ticket,

Also fix the override of kanban view with dashboard in midium screen.
also fix the width of the kanbancard  for xs device to avoid the
horizontal scrollbar.

closes odoo/enterprise#18272

Taskid: 2518829
Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>