[IMP] account_reports: tax closing with foreign vat: improve labelling
We now only display the country on the closing entry when it's useful, that is, when there are foreign VAT fiscal positions defined. Same goes for the states: we show them only when it adds information.

Also, we don't call the domestic closing entry 'Domestic' anymore. Instead, we label it with its country and country states when needed, just like any other closing move.

X-original-commit: c96c21ca5bdbe8e367fff4fe2e39675cd1fa3d59