[FIX] account_reports: tax closing: allow switching from 'annually' to 'monthly' without duplicating closing moves
Before, when switching from 'annually' to 'monthly' tax return periodicity with a draft closing move in the current period, a new move was created for the current month instead of overwriting the dates and labels of the existing one. This resulted in multiple closing next activities showing up on the dashboard.

Now, we overwrite the existing move properly.

closes odoo/enterprise#18236

X-original-commit: d21aa30f79bd35cfd7301267692fc57d3af3e5a3
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>