[FIX] account_reports: tax closing: properly handle warnings and empty tax closing
To reproduce the issue:

1) install a CoA without any default tax receivable of payable account, and no value set for these fields on any tax group (for example l10n_nl)
2) Post an invoice in a previous period
3) open the tax report for this previous period and click on "closing journal entry"
===> No error is raised, the entry is created with 2 lines at 0.

Now, try to click on the refresh buttonof the closing entry
===> A RedirectWarning is raised, as your tax groups aren't configured.

This is because of the way raise_on_empty is used since https://github.com/odoo/enterprise/commit/d4036da02b997b24115e0211d8709d565b1f2904 . When called from the report, it's False ; when called from the button, it's True.

We don't want this behavior. The check is now always performed: if the tax groups are not correctly configured, we raise. We also remove the second RedirectWarning, as it does not make sense anymore to raise in case of an empty closing move under any circumstances.

closes odoo/enterprise#18240

X-original-commit: 6e7eba4b79e5ca5c61d7f7961da6ecc05d8ab704
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>