[FIX] l10n_in_hr_payroll: fix the india payroll issue
Currently, the Employee and Worker structure type are common to all country payroll.
-When India payroll is installed, the default structure is set in this structure
type. But access rights warnings occur when the (India) country is empty on the company.
In this commit, the user has to set the default structure in the common structure type,
which will not be set by default.

-Non-Executive Employee structure will be set as default structure in
Non-Executive structure type.

Task-id:- 2524689

closes odoo/enterprise#18226

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>