[FIX] social: hide tour bubble message in modal
before this commit:
the tour bubble message "Write a message to get a preview
of your post" attaches itself to the comment field on the
post form view and in the modal opens from the kanban card.

after this commit:
the tour bubble message "Write a message to get a preview
of your post" attaches itself only to the comment field on the
post form view.

closes odoo/enterprise#18404

Task: 2525562
X-original-commit: 7c64158f2dccd3fa846bf1f54e4ba65117db4e28
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>