[FIX] account_accountant: compute balance of write-off move
When adding a write-off move to reconcile an invoice, its value will
always be the amount residual instead of the amount defined by the user.

To reproduce the error:
1. Accounting > Configuration > Reconciliation Models, Create
    - Type: Manually [...]
    - Add one Counterpart Values:
        - Account: Bank Fees
        - Amount Type: Fixed
        - Journal: Miscellaneous Operations
        - Amount: 50
2. Create/Confirm an invoice INV of $1000
3. Create a payment P
    - Type: Receive Money
    - Customer/Vendor: INV's customer
    - Amount: $600
4. Confirm P, Payment Matching
5. Select INV, P and add Bank Fees (Manual Operations)
6. Reconcile
7. Open INV

Error: The amount due is 0, there are two payments: one of $600 (payment
P) and a second one of $400. The latter is incorrect and should be $50
(bank fees).

The issue comes from method `_prepare_writeoff_moves`, its parameter
`vals` contains the following information:
```python
{
    'name': '1',
    'balance': -50,
    'analytic_tag_ids': [[6, None, []]],
    'account_id': 29,
    'journal_id': 3,
    'reconcile_model_id': 3
}
```
However, the method does not check if `vals` already has a balance
value. Instead, it computes this balance by adding up the residual
amounts. This is where the incorrect amount ($400) comes from.

OPW-2485656

closes odoo/enterprise#18302

X-original-commit: 211f74a7985cd0735adc61878d9020a6cf1f2c9a
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>
Signed-off-by: Adrien Widart <adwid@users.noreply.github.com>