[FIX] sale_amazon: fix sync of partially canceled orders
[FIX] sale_amazon: fix sync of partially canceled orders

Amazon allows customers to cancel only some items in an Amazon order
rather than canceling the entire order altogether. This results in
orders being synchronized with the quantity of some order lines being
set to 0. As the unit price is computed with `total_amount / quantity`,
a division by zero occurs.

This commit prevents such errors by setting the unit price to 0 if the
quantity is 0. This allows showing lines for canceled products like it
is done in SellerCentral.

opw-2523766

closes odoo/enterprise#18282

X-original-commit: 60457101597bbe9927ef3d051c016f71fade9979
Signed-off-by: Antoine Vandevenne (anv) <AntoineVDV@users.noreply.github.com>
Signed-off-by: Achraf <abz-odoo@users.noreply.github.com>