[FIX] pos_loyalty: Only show 3 control-buttons in product screen on mobile
Before this commit, when you clicked on "Review" the order lines weren't
visible due to the number of control-buttons.

To fix this, a "More..." button is displayed if there is more than 3 buttons.
When you click on this button, a new modal will display a menu with all controls.
Note that we had to find a way to close this popup when you triggered the action.

See the related commit in community for more information.

Task ID: 2453679

closes odoo/enterprise#22146

X-original-commit: 0cb088e7114ad53637778b8b995cd6131f24399f
Related: odoo/odoo#79448
Signed-off-by: Pierre Paridans (app) <app@odoo.com>
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>
Co-authored-by: Adrien Dieudonné <adr@odoo.com>