[FIX] l10n_lu_reports_electronic, l10n_lu_reports_electronic_xml_2_0: fix XML export 2.0 for 0% taxes in tax report
When exporting the tax report's monthly xml with enterprise, the file wasn't accepted by the government in case 0% taxes were used.

Grid 033 is optional, and can be used for custom tax rates. If specified, it MUST always go with grid 403 (giving the tax rate) and 042 (giving the tax amount). In Odoo, we made the choice to use it for 0% taxes, so 403 and 42 will always be present, and always 0.

In enterprise, 403 grid needed to always be injected to the XML values, and grids 033, 042 and 0433 had to be made mandatory in the xml, so that they are not removed from it when amounting to 0.

closes odoo/enterprise#18305

X-original-commit: c50e97efff3072e9e85271135932c7d94cce55aa
Related: odoo/odoo#70792
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>