[MOV] web_enterprise, web_mobile: move files to the right location
Before this commit, the customisation for the pivot view was made on
'Mobile', but the files was placed in the wrong directory
'web_mobile'.

After this commit, these files are moved in 'web_enterprise'.
web_mobile should be only used for native features of the mobile app.

Part-of: odoo/enterprise#18121