[FIX] *: update documentation links
Following the recent reorganisation of the documentation in 12.0+, the
majority
of the documents have been moved and their old links are no longer
valid.

Some redirection rules will soon be deployed, but those rules might be
dropped in some
years and we want the links to still work, which is why we still replace
the links to the
new ones.

FW-port of odoo/enterprise#18270 (13.0)

closes odoo/enterprise#18361

X-original-commit: eba1f3bc427ae4381e675798de019a5d012c4ab3
Related: odoo/odoo#70920
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>