[FIX] project_(timesheet_)grid,synchro: improve the user/validation errors
Purpose of the commit is, User/validation errors are sometimes not
correct in English and/or quite obscure and don't help the user
understand/solve the issue. so some copywriting should help make
the experience better.

So in this commit, re-word the several validation and usererror message
and also made a user with the 'timesheets > see own timesheets' access
right level shouldn't have access to menus other than "My Timesheets"

Also disable the edition of the description field of validated
timesheets for users with the 'see own timesheets' access right level

also while validating the timesheet if there is timer start on analytic
line then stop the line and validate it instead of raising the notification.

Related Community PR: https://github.com/odoo/enterprise/pull/18386

closes odoo/enterprise#18386

Taskid: 2513067
Related: odoo/odoo#70796
Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>