[FIX] l10n_ar_edi: QR code for Exportation Invoices
We found out an error when creating the QR code for the PDF
representation of the "(19) Factura de Exportacion" document type. We
check an AFIP Online generated invoice and they use CUIT PAIS as the
receptor document number.

Formerly we were reporting a sanitized version of the VAT number of the
foreign customer because QR code specification tells us that the
receptor document number is a numeric value, which was causing confusion
because this one may be alphanumeric and we were sending a modified
fiscal number.

Now we are using the CUIT PAIS instead which is a generic CUIT number
defined by AFIP to Identify receptors for a specific country, This way
we are ensuring that the QR receptor document number is numeric.

closes odoo/enterprise#21707

X-original-commit: 890f68e2e31c9f98150607a6e0d49520fb57d46f
Signed-off-by: Josse Colpaert <jco@openerp.com>