[MOV] documents_spreadsheet: Move autofill in its own plugin
Currently, the `PivotStructure` plugin is 1.2k LoC in which it's not easy
to navigate.

One can see that the plugin has two responsabilities:
- pivot structure
- pivot autofill

Those responsabilities are independent, no attribute nor method are shared.
This commit moves all autofill logic in a separate plugin.
This makes two smaller plugins which are more focused and easier to hold in
one's head.