[FIX] stock_enterprise: properly calculates dates in transfers
In e.g. 12.0 when transferring inventory, if the client's timezone is not UTC, the delay and cycle time can be wrong if there is one date that is not on the same day for UTC and the client's timezone.

With this commit, we allow decimals in dates instead of truncating to calculate only on number of days, and round the result properly.

opw-2470072

closes odoo/enterprise#18326

X-original-commit: 9ed2c44ed5937baa66fc0bab186a60d53c18de7b
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>
Signed-off-by: pvh-odoo <SwagSamaSempai@users.noreply.github.com>