[IMP] planning,project_forecast: remove company field in shift template
Before this commit, the shift template is linked to a company, by
default the current one. But we could need to the same shift in the
other companies.

Rather than duplicate the shift template, this commit removes the
company field in `planning.slot.template` model and add
company_dependant=True in project_id and task_id fields in this same
model to keep the company dependant for those fields.

task-2463722

Upgrade PR: odoo/upgrade#2426

closes #17958