[FIX] stock_barcode: create picking after refresh
In the Barcode App, when the user creates a new picking with the
"Create" button, an action will be returned if the active model is
`stock.picking.type` in the context.
But as the context is lost when the user refreshes the browser, no
action is returned but it will try to do an action anyway, so it will
get a traceback.

How to reproduce:
 - Open Barcode;
 - Go to "Operations" and select an operation type;
 - Refresh (F5) then click on "Create" button -> Traceback.

With this fix, it will open the form view to create a new picking, which
isn't a perfect solution but still better than a traceback.

closes odoo/enterprise#18521

X-original-commit: d96c8ffcaa01ca6a93af97a3583b72afd458b036
Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>
Signed-off-by: Steve Van Essche <svs-odoo@users.noreply.github.com>