[FIX] sign: remove fields' name on PDF viewer
When consulting a signed document, if the user didn't complete some
optional fields, theses ones are still displayed with their placeholder
or field name as value

Note: This issue only happens on PDF preview. The downloaded PDF does
not contain any undesirable values.

To reproduce the issue:
1. In Sign, upload a PDF Template:
    - Add two fields:
        - Name
        - Text ("Mandatory field" disabled)
2. Sign Now
3. Complete Name, but not Text
4. Validate & Send Completed Document
5. Sign > Signatures, open signed document

Error: in the PDF viewer, the field Text is present on the document
(with value "Text"). It should not be present (The PDF provided by
"Download Document" button does not contain undesirable fields)

When the PDF is fully signed, a new document is generated and unfilled
fields are ignored. See in method `generate_completed_document`:
https://github.com/odoo/enterprise/blob/318da49017efbe0f2f9a045e022f601725d47f26/sign/models/sign_request.py#L468-L471
However, nothing prevents these fields to be rendered in the Sign
application viewer.

OPW-2479314

closes odoo/enterprise#18480

X-original-commit: a3f848ec99f695e04d5c09479886cd26ba3dc68d
Signed-off-by: Adrien Widart <adwid@users.noreply.github.com>