[FIX] l10n_cl_edi_stock: make option 'Batch Pickings' possible
Error: When module `l10n_cl_edi_stock` is installed, if the user tries
to enable the option 'Batch Pickings', it will lead to an Odoo Error
with a traceback.

When the module `stock_picking_batch` is installed (i.e., when 'Batch
Picking' is enabled), a view is created based on
`stock.view_picking_form`, except that the header is replaced:
https://github.com/odoo/odoo/blob/b4e1e9f3ae86e6b5e7015c0241791f733ee942f4/addons/stock_picking_batch/views/stock_picking_batch_views.xml#L10-L13
However, in `l10n_cl_edi_stock`, the field `country_code` is declared in
the replaced header. As a result, (still in `l10n_cl_edi_stock`) the
invisibility condition can not be computed:
https://github.com/odoo/enterprise/blob/6cd63c5c64073e38e4b90ea612a0928c23983df2/l10n_cl_edi_stock/views/stock_picking_views.xml#L35

OPW-2487757

closes odoo/enterprise#18574

X-original-commit: 2d0500556a2ba596977dd64a50f7a8b94823c036
Signed-off-by: Adrien Widart <adwid@users.noreply.github.com>