[FIX] website_calendar: fix the design of slots selection
Improve the design of the page where we select a slot to take an
appointment. We remove the scrollbar present for the displaying the
slots list. Now we display the slots with 2 columns and if there are
a lot of slots, we increase the heigth of the page.
For the mobile view, we dipslay only one column.

task-2536019

closes odoo/enterprise#18590

X-original-commit: c9b570440d4894ab8db21405e86a8a21fb3b2eef
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>