[FIX] account_sepa_direct_debit: manage payments using invoice contact
In case the partner set on the payment is of type invoice, the name is not mandatory
and is generally not set. This is causing a traceback when generating the SDD file
as the field Debtor Name must be filled in. We should add a default case and take the
parent company name instead.

closes odoo/enterprise#18487

Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>
Signed-off-by: Alex Tuyls <alt-odoo@users.noreply.github.com>