[fix] account_intrastat : MemoryError on install when lots of move lines
Installing 'account_intrastat' on a database with a lot of 'account.move.line'
could lead to a 'MemoryError', in addition to take a lot of time (for no reason)
when the ORM was computing the fields
'intrastat_product_origin_country_id'

which are related, stored, fields from `product.template`,
which are empty/`NULL` on install (no default set for the field),
and therefore the ORM was taking a lot of time and memory
for, at the end, setting `NULL` for all records.

By setting the columns in advance,
the ORM doesn't attempt to compute these fields
on install.

opw-2532754

closes odoo/enterprise#18552

X-original-commit: 1078e1259b1bb33c9d9884dd5da577607970b27b
Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>