[FIX] account_asset: hides Existing Depreciation Schedule in Models
In e.g. 14.0 when in debug mode, the Existing Deprecation Schedule group of fields shows in Deferred Revenue Models, when it should only be visible on Deferred Revenues themselves.

With this commit, we hide the Existing Deprecation Schedule group of fields in the case where the state (status) of the current Deferred Revenue is set to model.

opw-2474486

closes odoo/enterprise#18585

X-original-commit: feee7f1f73c6c25f56072418e4e004328c8dc7a5
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>
Signed-off-by: pvh-odoo <SwagSamaSempai@users.noreply.github.com>