[FIX] account_reports: Fix analytic report filters since 1f325e84740aac6a686cd6df4cf9b9ff5d4edc8e
filter_analytic_accounts/filter_analytic_tags are declared on 'self' only when using financial reports.
If we tried to open a report being not a financial report but using the analytic filters, it leads to a KeyError: 'analytic_accounts'.
This is because '_init_filter_analytic_accounts' is not executed since there is no existing filter 'filter_analytic_accounts' but 'filter_analytic'.

Introduced by:
https://github.com/odoo/enterprise/commit/1f325e84740aac6a686cd6df4cf9b9ff5d4edc8e

closes odoo/enterprise#18592

Co-author: adwid <awt@odoo.com>
X-original-commit: 1d5dc159d28434a86d09927c01d573e1ddeefdaf
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>