[FIX] l10n_ar_edi: dummy local invoice validation
We were not able to do dummy local validation of electronic invoices. We
found that the problem was related to the order that the account.move is
processed:

1. we try to post the AR electronic invoice in Odoo first.
2. if everything is ok then we validate it in AFIP.

Since in the Odoo the invoice post() method is computing the invoice
name then we are also trying to connect to AFIP to get the last invoice
number that will be used to set this value.

Adding a call to _is_dummy_afip_validation() method in the method that
gets the last sequence of the invoice to resolve the problem.

closes odoo/enterprise#18781

X-original-commit: 6df1d4f99c0c4c403891df09886fd5ca950ec2a9
Signed-off-by: Josse Colpaert <jco@openerp.com>