[IMP] various: beautify list-views using decoration
Currently, to identify records based on state, in most of the list views we
use 'badge' widget on 'state' fields which can be decorated as needed.

However, for the list views, best approach would be to also add list
decoration along with 'badge' widget so that the whole line is colored
and it's easy for user to 'scan' the information.

This commit improves the behavior adding decorations to several list-views.
To check the list of changes, kindly see the task specification

Task-ID:2527119

closes odoo/enterprise#18587

Related: odoo/odoo#71367
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>