[FIX] timesheet_grid: change the position of the overtime indication
Before this commit, the Many2OneAvatar widget has recently been changed
to add a div in its template. This changes impacted the
TimesheetM2OAvatarEmployee widget in timesheet_grid module because this
widget uses Many2OneAvatar widget in intern and then the positioning of
the elements in the TimesheetM2OAvatarEmployee widget has changed.

This commit changes the position of the overtime indication we add it in
the new div created in the Many2OneAvatar to keep the initial behaviour
and alignment.

Related PR: odoo/odoo#67366

closes odoo/enterprise#18605

Related: odoo/odoo#71461
Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>