[IMP] account_reports: seperate column fields into Monetary and Float
Accounting reports use the same formatting for float and monetary fields
This commit seperates the field types and formats them differently. We also add a parameter blank_if_zero to replicate the behaviour of using an empty string when the value is 0.00
Also add a field_column for fields of type many2one
We adapt the multicurrency revaluation wizard for this change.

in preparation for task 2352602