[MOV] appointment: make appointment app for portal
With this commit, the appointment functionalities no longer
need the use of the website module. We take this opportunity
to name this app "appointment". It shows more clearly the purpose
of the app which is to manage appointment.

Use case:
e.g. We have a Dental Office that don't need a website but still want
to send appointment links to its patients. The patient will get the
link and follow the usual process to take the appointment.
The user will no longer need to buy the website module to use these
features.

The appointment app will now depends of the portal module. The website
features that were present are now available in the website_appointment
module.

These features are:
	- The /calendar route and linked templates (list and cards layout)
	- The management of website visitor in the process of taking
an appointment
	- The website related mixins

The purpose to keep the /calendar route in the website version is to
have a minimum of security. The appointment_type are automatically now
published without the website module. The goal is to avoid the public
to have access to all the possible appointment types easily.

We avoid here to have to much of security layers to match the use cases
of this app. In the case, someone know the base url it will still be
possible to get all appointment types available and in some case the name
of the employees asigned to one. But the process of getting them is hard
enough to avoid other security layers.

Additionnaly, the module "website_calendar_crm" is renamed
"appointment_crm". We don't need the website to use the crm part.

task-2427015
COM PR: odoo/odoo#72241
ENT PR: odoo/enterprise#18426
UPG PR: odoo/upgrade#2492