[IMP] planning: add active/archive feature in planning.role
Before this commit, if the user wants to hide a role in differents views
using planning.role model, he cannot because he cannot archive the
intances of this model.

This commit adds the active/archive feature to allow the user to archive
the role that he wants to archive.

task-2463722

Upgrade PR: odoo/upgrade#2426

closes #17958