[FIX] website_twitter_wall: update deprecated Twitter route
This commit updates the twitter route URL used to fetch the tweet HTML.

Indeed, the old "oembed" route is deprecated and will be removed completely
soon (November 23, 2021).
More information here:
https://twittercommunity.com/t/consolidating-the-oembed-functionality/154690

In order to use the updated oembed route, we have to pass an additional
author_id parameter, already contained in the retrieved tweet information.
This implies a modification of the method signature, so the old route was kept
for retro-compatibility in case the parameter is missing.

Task-2544931

closes odoo/enterprise#18663

X-original-commit: a80a17f18b17a699c4e6b59118ed14410216735a
Signed-off-by: awa-odoo <awa-odoo@users.noreply.github.com>
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>