[ADD] sale_amazon_spapi: migrate from the MWS API to the SP-API
Starting July 2022, Amazon will disable some endpoints of the MWS API
in favor of the new Selling Partners API (SP-API). This commit adds a
new module `sale_amazon_spapi` to patch the business methods of the
currently existing modules of the Amazon Connector and interface with
the SP-API.

Some changes are made to the base modules too but these should still
function normally until Amazon disables the MWS API or the new module is
installed.

task-2466636

closes odoo/enterprise#27939

X-original-commit: 493515625e3533b6e7d604acb5e3ff7e8f359267
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>
Co-authored-by: Antoine Vandevenne (anv) <anv@odoo.com>