[FIX] account_reports: fix tax closing from the generic tax report
Generic views from the tax report are not related to any country. This caused the tax closing code to disregard foreign VAT fiscal positions, and only do the closing for the domestic country when the button was clicked from a generiv report. What we want instead is the closing to be done for ALL fiscal positions available for this company, whatever their country.

closes odoo/enterprise#18602

X-original-commit: 01187548851cd5acf2a812e4517ba8fa0b36ff6b
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>