[FIX] mass_mailing_themes: properly set new background colors in columns
Colorpicker snippets target <td> elements while in the case of two and
three column snippets, these contain a <div> which is background-colored
in css. In consequence, when using the colorpicker to change the
background color of the snippet, the colored <div> came in front of the
element of which we tried to change the color. This ensures the snippets
target the <div> instead as this is then in fact the element we want to
color.

closes odoo/enterprise#18619

Related: odoo/odoo#71500
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>