[FIX] document_spreadsheet: manage total LR autofill
Steps:
- Install Accounting
- Go to Accounting / Reporting / Management / Budgets Analysis
- Switch to Pivot view
- Click Insert in Spreadsheet
- Confirm
- Try to autofill the Total row header

Bug:
Traceback here:
https://github.com/odoo/enterprise/blob/72541b4904adc028e4700ff9e5be93e90ed7907e/documents_spreadsheet/static/src/js/o-spreadsheet/plugins/pivot_plugin.js#L1081
cols is undefined

Explanation:
`cols` is undefined because the Total header autofill goes through
`_autofillPivotColHeader` while autofilling a row header.
Total headers don't have arguments. This makes them fit neither in row
headers nor in column headers.

This commit uses the direction of the autofill to know if the Total
header is a row or a column header.

opw:2522672

closes odoo/enterprise#18691

X-original-commit: 2550490879f3839cd2591737464f7bcd3ea9af25
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>
Signed-off-by: backspac <backspac@users.noreply.github.com>