[FIX] social_facebook: update endpoints to latest API
The social_facebook module is still using old API (v3.3) endpoints that are
deprecated and will soon not be supported anymore by Facebook.

This commit updates the various endpoints to use the latest version (10.0).

Most of the changes just imply updating the URL from 3.3 to 10.0.
Here are a couple exceptions:
- The permissions were reviewed in API v7.0 and were split for more granularity
  See https://developers.facebook.com/docs/graph-api/changelog/version7.0/
  This doesn't change much for us, we just ask the new permissions names
  instead.
- The '/photos' endpoint incorrectly saved the facebook_post_id

Task-2551155
IAP PR odoo/iap-apps#352

closes odoo/enterprise#18771

X-original-commit: 0faa5d57febc7430169834a3746733850e968c07
Signed-off-by: awa-odoo <awa-odoo@users.noreply.github.com>
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>