[FIX] web_studio: prevent infinite loop if studio bundle crashes
Before this commit, if there was a syntax error in the studio lazy
loaded bundle, that bundle was loaded over and over until the browser
died. The reason is that the studio client action is initially a
function that loads the whole bundle. In that bundle, we override the
client action in the registry with the real one. The initial function
returns the same action, s.t. when it is done (when the bundle is
loaded), we execute again the studio action (which is supposed to be
the real action now). If there is a crash in the lazy loaded bundle,
the action could not be overriden in the registry, meaning that when
the action is executed again, it is still the function to lazy load
the bundle, and so on.

This commit ensures that we directly replace the action in the
registry by another one, s.t. whatever happens, the bundles are
only loaded once.