[IMP] social: allow multiple companies to use Social
Purpose
=======
Allow multiple companies to use Social on the same database.

So each company does not want to share their social accounts with the
others companies.

Specifications
==============
All the social records (post, account...) belong to one company or none.
For the social post, if one company is set, we can only select account
in the same company or without a company. If no company is set on the
post, we can select any social accounts.

Technical
=========
During the authentication process, the user is redirected to the social
media website (Facebook, LinkedIn...) and then to a Odoo endpoint.

When the user is redirected to the Odoo endpoint, the selected company
on the backend is not preserved and `self.env.company` always return
the company of the current user and not the selected company on the
backend like it does for standard action.

To be able to give the company of the new accounts, we store it in the
user session (so we should not pass a parameter for all social media
authentication, which would require custom code for each media).

Task 2415987
See odoo/enterprise/pull/15896
See odoo/upgrade/pull/2099